--- packages/shared/src/components/styles/Modal.css
@@ -1,132 +1,0 @@
-@layer components {
-  /* ========== Animations ========== */
-  @keyframes modal-fade {
-    from {
-      opacity: 0;
-    }
-
-    to {
-      opacity: 1;
-    }
-  }
-
-  @keyframes modal-zoom {
-    from {
-      transform: scale(0.8);
-    }
-
-    to {
-      transform: scale(1);
-    }
-  }
-
-  @keyframes mymodal-blur {
-    from {
-      background: rgba(45 0 0 / 0);
-      backdrop-filter: blur(0);
-    }
-
-    to {
-      background: rgba(45 0 0 / 0.3);
-      backdrop-filter: blur(10px);
-    }
-  }
-
-  @keyframes mymodal-slide {
-    from {
-      transform: translateX(100%);
-    }
-
-    to {
-      transform: translateX(0);
-    }
-  }
-
-  /* ========== Modal Overlay ========== */
-  .react-aria-ModalOverlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100vw;
-    height: var(--visual-viewport-height);
-    background: color-mix(in srgb, var(--fg) 32%, transparent);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    z-index: 100;
-
-    &[data-entering] {
-      animation: modal-fade 200ms;
-    }
-
-    &[data-exiting] {
-      animation: modal-fade 150ms reverse ease-in;
-    }
-  }
-
-  /* ========== Base Modal Styles ========== */
-  .react-aria-Modal {
-    box-shadow: 0 8px 24px
-      color-mix(in srgb, var(--fg) 15%, transparent);
-    border-radius: var(--radius-xl);
-    background: var(--bg-overlay);
-    color: var(--fg);
-    border: 1px solid var(--border);
-    outline: none;
-    max-width: 300px;
-
-    &[data-entering] {
-      animation: modal-zoom 300ms cubic-bezier(0.175, 0.885, 0.32, 1.275);
-    }
-
-    .react-aria-TextField {
-      margin-bottom: var(--spacing);
-    }
-  }
-
-  /* ========== Color Variants ========== */
-  .react-aria-Modal[data-variant="primary"] {
-    border-color: var(--accent);
-  }
-
-  .react-aria-Modal[data-variant="secondary"] {
-    border-color: var(--border);
-  }
-
-  .react-aria-Modal[data-variant="tertiary"] {
-    border-color: var(--color-purple-600);
-  }
-
-  .react-aria-Modal[data-variant="error"] {
-    border-color: var(--negative);
-  }
-
-  /* ===== Filled Variant ===== */
-  .react-aria-Modal[data-variant="filled"] {
-    background: var(--accent-subtle);
-    border: none;
-  }
-
-  /* ========== Size Variants ========== */
-  .react-aria-Modal[data-size="sm"] {
-    max-width: 240px;
-    border-radius: var(--radius-lg);
-  }
-
-  .react-aria-Modal[data-size="md"] {
-    max-width: 300px;
-    border-radius: var(--radius-xl);
-  }
-
-  .react-aria-Modal[data-size="lg"] {
-    max-width: 400px;
-    border-radius: var(--radius-2xl);
-  }
-
-  /* ========== Forced Colors Mode (High Contrast) ========== */
-  @media (forced-colors: active) {
-    .react-aria-Modal {
-      forced-color-adjust: auto;
-    }
-  }
-}
