--- packages/shared/src/components/styles/Tabs.css
@@ -1,315 +1,0 @@
-@layer components {
-  /* ========== Base Tabs Container ========== */
-  .react-aria-Tabs {
-    width: 100%;
-    display: flex;
-    color: var(--fg);
-
-    /* 기본 variant 변수 */
-    --tab-accent: var(--accent);
-
-    &[data-orientation="horizontal"] {
-      flex-direction: column;
-    }
-
-    &[data-orientation="vertical"] {
-      flex-direction: row;
-    }
-  }
-
-  /* ========== Base TabList ========== */
-  .react-aria-TabList {
-    width: 100%;
-    display: flex;
-    position: relative;
-
-    /* 기본 variant 변수 */
-    --tab-accent: var(--accent);
-
-    &[data-orientation="horizontal"] {
-      border-bottom: 1px solid var(--border);
-    }
-
-    &[data-orientation="vertical"] {
-      flex-direction: column;
-      border-inline-end: 1px solid var(--border);
-    }
-  }
-
-  /* ========== React Aria 1.13.0: SelectionIndicator ========== */
-  .react-aria-Tab .react-aria-SelectionIndicator {
-    position: absolute;
-    background: var(--tab-accent);
-    border-radius: 0;
-    pointer-events: none;
-    transition:
-      translate 200ms cubic-bezier(0.16, 1, 0.3, 1),
-      width 200ms cubic-bezier(0.16, 1, 0.3, 1),
-      height 200ms cubic-bezier(0.16, 1, 0.3, 1);
-
-    @media (prefers-reduced-motion: reduce) {
-      transition: none;
-    }
-  }
-
-  .react-aria-TabList[data-orientation="horizontal"]
-    .react-aria-Tab
-    .react-aria-SelectionIndicator {
-    left: 0;
-    bottom: 0;
-    width: 100%;
-    height: 3px;
-  }
-
-  .react-aria-TabList[data-orientation="vertical"]
-    .react-aria-Tab
-    .react-aria-SelectionIndicator {
-    top: 0;
-    right: 0;
-    height: 100%;
-    width: 3px;
-  }
-
-  .react-aria-TabList[data-show-indicator="true"] .react-aria-Tab::before {
-    display: none;
-  }
-
-  /* ========== Base Tab ========== */
-  .react-aria-Tab {
-    position: relative;
-    padding: var(--spacing-xs) var(--spacing-md);
-    font-size: var(--text-sm);
-    font-weight: 400;
-    cursor: pointer;
-    outline: none;
-    color: var(--fg-muted);
-    background: transparent;
-    border: none;
-    transition:
-      color 200ms,
-      background 200ms;
-    forced-color-adjust: none;
-
-    &[data-hovered] {
-      color: var(--fg);
-      background: color-mix(in srgb, var(--fg) 8%, transparent);
-    }
-
-    &[data-focused] {
-      color: var(--fg);
-    }
-
-    &[data-selected] {
-      color: var(--fg);
-      font-weight: 600;
-    }
-
-    &[data-disabled] {
-      color: color-mix(in srgb, var(--fg) 38%, transparent);
-      cursor: not-allowed;
-      pointer-events: none;
-    }
-
-    &[data-focus-visible]::after {
-      content: "";
-      position: absolute;
-      inset: 4px;
-      border-radius: var(--radius-sm);
-      outline: 2px solid var(--tab-accent);
-      outline-offset: 2px;
-    }
-
-    &[data-orientation="horizontal"]::before {
-      content: "";
-      position: absolute;
-      bottom: 0;
-      left: 0;
-      right: 0;
-      height: 3px;
-      background: transparent;
-      transition: background 200ms;
-    }
-
-    &[data-orientation="horizontal"][data-selected]::before {
-      background: var(--tab-accent);
-    }
-
-    &[data-orientation="vertical"]::before {
-      content: "";
-      position: absolute;
-      top: 0;
-      bottom: 0;
-      right: 0;
-      width: 3px;
-      background: transparent;
-      transition: background 200ms;
-    }
-
-    &[data-orientation="vertical"][data-selected]::before {
-      background: var(--tab-accent);
-    }
-
-    &[href] {
-      text-decoration: none;
-      cursor: pointer;
-    }
-  }
-
-  /* ========== Base TabPanel ========== */
-  .react-aria-TabPanel {
-    width: 100%;
-    padding: var(--spacing-lg);
-    border-radius: var(--radius-sm);
-    outline: none;
-    color: var(--fg);
-
-    &[data-focus-visible] {
-      outline: 2px solid var(--accent);
-      outline-offset: 2px;
-    }
-  }
-
-  /* ========== Density Variants ========== */
-
-  /* compact: 좁은 간격 */
-  .react-aria-Tabs[data-density="compact"],
-  .react-aria-TabList[data-density="compact"] {
-    .react-aria-Tab {
-      padding: var(--spacing-2xs) var(--spacing-sm);
-      font-size: var(--text-sm);
-    }
-
-    .react-aria-TabPanel {
-      padding: var(--spacing-md);
-    }
-  }
-
-  /* regular (기본): 기본 간격 — 명시적 선언 */
-  .react-aria-Tabs[data-density="regular"],
-  .react-aria-TabList[data-density="regular"] {
-    .react-aria-Tab {
-      padding: var(--spacing-xs) var(--spacing-md);
-      font-size: var(--text-sm);
-    }
-
-    .react-aria-TabPanel {
-      padding: var(--spacing-lg);
-    }
-  }
-
-  /* ========== Size Variants ========== */
-  .react-aria-Tabs[data-size="sm"],
-  .react-aria-TabList[data-size="sm"] {
-    .react-aria-Tab {
-      padding: var(--spacing-2xs) var(--spacing-sm);
-      font-size: var(--text-xs);
-
-      &[data-orientation="horizontal"]::before,
-      &[data-orientation="vertical"]::before {
-        width: 2px;
-        height: 2px;
-      }
-    }
-
-    .react-aria-TabPanel {
-      padding: var(--spacing-sm);
-    }
-
-    &[data-orientation="horizontal"]
-      .react-aria-Tab
-      .react-aria-SelectionIndicator {
-      height: 2px;
-    }
-
-    &[data-orientation="vertical"]
-      .react-aria-Tab
-      .react-aria-SelectionIndicator {
-      width: 2px;
-    }
-  }
-
-  .react-aria-Tabs[data-size="md"],
-  .react-aria-TabList[data-size="md"] {
-    .react-aria-Tab {
-      padding: var(--spacing-xs) var(--spacing-md);
-      font-size: var(--text-sm);
-      line-height: var(--text-sm--line-height);
-
-      &[data-orientation="horizontal"]::before {
-        height: 3px;
-      }
-
-      &[data-orientation="vertical"]::before {
-        width: 3px;
-      }
-    }
-
-    .react-aria-TabPanel {
-      padding: var(--spacing-md);
-    }
-
-    &[data-orientation="horizontal"]
-      .react-aria-Tab
-      .react-aria-SelectionIndicator {
-      height: 3px;
-    }
-
-    &[data-orientation="vertical"]
-      .react-aria-Tab
-      .react-aria-SelectionIndicator {
-      width: 3px;
-    }
-  }
-
-  .react-aria-Tabs[data-size="lg"],
-  .react-aria-TabList[data-size="lg"] {
-    .react-aria-Tab {
-      padding: var(--spacing-sm) var(--spacing-lg);
-      font-size: var(--text-base);
-
-      &[data-orientation="horizontal"]::before {
-        height: 4px;
-      }
-
-      &[data-orientation="vertical"]::before {
-        width: 4px;
-      }
-    }
-
-    .react-aria-TabPanel {
-      padding: var(--spacing-lg);
-    }
-
-    &[data-orientation="horizontal"]
-      .react-aria-Tab
-      .react-aria-SelectionIndicator {
-      height: 4px;
-    }
-
-    &[data-orientation="vertical"]
-      .react-aria-Tab
-      .react-aria-SelectionIndicator {
-      width: 4px;
-    }
-  }
-
-  /* ========== Forced Colors Mode (High Contrast) ========== */
-  @media (forced-colors: active) {
-    .react-aria-Tabs,
-    .react-aria-TabList,
-    .react-aria-Tab,
-    .react-aria-TabPanel {
-      forced-color-adjust: auto;
-    }
-
-    .react-aria-Tab {
-      &[data-selected]::before {
-        background: Highlight;
-      }
-
-      &[data-focus-visible]::after {
-        outline-color: Highlight;
-      }
-    }
-  }
-}
