--- packages/shared/src/components/styles/ToggleButtonGroup.css
@@ -1,161 +1,0 @@
-@layer components {
-  .react-aria-ToggleButtonGroup {
-    display: flex;
-
-    &:focus-within {
-      outline: 2px solid var(--focus-ring);
-      outline-offset: 2px;
-      border-radius: var(--radius-lg);
-    }
-
-    > button {
-      border-radius: 0;
-      z-index: 1;
-
-      &[data-disabled] {
-        z-index: 0;
-      }
-
-      &[data-selected] {
-        z-index: 2;
-      }
-
-      &[data-focus-visible] {
-        z-index: 2;
-        outline: none;
-      }
-    }
-
-    /* ===== Indicator Mode ===== */
-
-    &[data-indicator="true"] {
-      /* Local color variable — variant overrides only this */
-      --indicator-bg: var(--bg-overlay);
-      --indicator-text: var(--fg-on-accent);
-
-      .react-aria-ToggleButton .react-aria-SelectionIndicator {
-        position: absolute;
-        inset: 0;
-        z-index: -1;
-        border-radius: var(--radius-sm);
-        background: var(--indicator-bg);
-        box-shadow: var(--shadow-sm);
-        pointer-events: none;
-        transition:
-          translate 200ms cubic-bezier(0.16, 1, 0.3, 1),
-          width 200ms cubic-bezier(0.16, 1, 0.3, 1),
-          height 200ms cubic-bezier(0.16, 1, 0.3, 1);
-
-        @media (prefers-reduced-motion: reduce) {
-          transition: none;
-        }
-      }
-
-      /* Variant overrides for indicator background */
-      &[data-togglebutton-variant="primary"]
-        .react-aria-ToggleButton
-        .react-aria-SelectionIndicator {
-        background: var(--accent);
-      }
-
-      &[data-togglebutton-variant="secondary"]
-        .react-aria-ToggleButton
-        .react-aria-SelectionIndicator {
-        background: var(--bg-inset);
-      }
-
-      &[data-togglebutton-variant="tertiary"]
-        .react-aria-ToggleButton
-        .react-aria-SelectionIndicator {
-        background: var(--color-purple-600);
-      }
-
-      &[data-togglebutton-variant="error"]
-        .react-aria-ToggleButton
-        .react-aria-SelectionIndicator {
-        background: var(--negative);
-      }
-
-      &[data-togglebutton-variant="surface"]
-        .react-aria-ToggleButton
-        .react-aria-SelectionIndicator {
-        background: var(--accent-subtle);
-      }
-
-      .react-aria-ToggleButton {
-        position: relative;
-        background-color: transparent;
-        border-width: 0;
-
-        &[data-pressed] {
-          background: transparent;
-          color: var(--fg);
-        }
-
-        &[data-selected] {
-          background: transparent;
-          color: var(--indicator-text);
-
-          &[data-pressed] {
-            background: transparent;
-            color: var(--fg);
-          }
-        }
-      }
-
-      /* Variant overrides for selected button text */
-      &[data-togglebutton-variant="secondary"]
-        .react-aria-ToggleButton[data-selected],
-      &[data-togglebutton-variant="tertiary"]
-        .react-aria-ToggleButton[data-selected],
-      &[data-togglebutton-variant="error"]
-        .react-aria-ToggleButton[data-selected] {
-        color: var(--color-white);
-      }
-
-      &[data-togglebutton-variant="surface"]
-        .react-aria-ToggleButton[data-selected] {
-        color: var(--fg);
-      }
-    }
-  }
-
-  /* ===== Orientation ===== */
-
-  .react-aria-ToggleButtonGroup[data-orientation="horizontal"] {
-    flex-direction: row;
-    width: fit-content;
-
-    > button {
-      margin-inline-start: -1px;
-
-      &:first-child {
-        border-radius: var(--radius-lg) 0 0 var(--radius-lg);
-        margin-inline-start: 0;
-      }
-
-      &:last-child {
-        border-radius: 0 var(--radius-lg) var(--radius-lg) 0;
-      }
-    }
-  }
-
-  .react-aria-ToggleButtonGroup[data-orientation="vertical"] {
-    flex-direction: column;
-    width: fit-content;
-
-    > button {
-      margin-block-start: -1px;
-      width: 100%;
-
-      &:first-child {
-        border-radius: var(--radius-lg) var(--radius-lg) 0 0;
-        margin-block-start: 0;
-      }
-
-      &:last-child {
-        border-radius: 0 0 var(--radius-lg) var(--radius-lg);
-      }
-    }
-  }
-}
