+++ packages/specs/src/components/Tabs.spec.ts
@@ -40,7 +40,7 @@ export const TabsSpec: ComponentSpec<TabsProps> = {
   name: "Tabs",
   description: "React Aria 기반 탭 컴포넌트",
   element: "div",
-  skipCSSGeneration: true,
+  skipCSSGeneration: false,
 
   defaultVariant: "default",
   defaultSize: "md",
